feat: add commands blueprint and implement commands route
