+++ abbybot_website/app/__init__.py
@@ -38,6 +38,7 @@ def create_app():
     from app.routes.status_data import status_data_bp
     from app.routes.bot_policies import bot_policies_bp
     from app.routes.user_responsibilities import user_responsibilities_bp
+    from app.routes.commands import commands_bp
 
     from app.routes.handlers.error_handlers import error_handlers_bp
 
@@ -52,5 +53,6 @@ def create_app():
     app.register_blueprint(status_data_bp)
     app.register_blueprint(bot_policies_bp)
     app.register_blueprint(user_responsibilities_bp)
+    app.register_blueprint(commands_bp)
     
     return app
